Moved the tiff_from_bbox classmethod to GrabCutDashboard
